+++ python_packages/habana_frameworks/torch/jit/csrc/jit_fork/ir/ir.cpp
@@ -1821,7 +1821,7 @@ Node* Graph::createTupleSlice(
   int64_t i = beg;
   for (const auto j : c10::irange(num_values)) {
     (void)j; // Suppress unused variable warning
-    auto idx = insertConstant(IValue(static_cast<int64_t>(i)));
+    auto idx = insertConstant(IValue(i));
     auto tupleIndex = insertNode(createTupleIndex(tup, idx, tt->elements()[i]));
 
     new_vals.push_back(tupleIndex->output());
